Allow for deployment of Nexus Professional
Add in the ability to deploy a Nexus Professional system by just setting
a single boolean (default false: aka deploy Nexus OSS) to true.

This adds in base rspec tests for the nexus::package class as well as
tests needed to validate that setting deploy_pro does the correct thing.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>